block: change size type to size_t
Use size_t instead of unsigned int, as used by strncpy().

+++ block.c
@@ -100,7 +100,7 @@ child_exec(struct block *block)
 }
 
 static void
-linecpy(char **lines, char *dest, unsigned int size)
+linecpy(char **lines, char *dest, size_t size)
 {
 	char *newline = strchr(*lines, '\n');
